+++ client/src/components/shop/index.js
@@ -1,0 +1,8 @@
+export { default as PriceSearch } from './PriceSearch'
+export { default as Menus } from './Menus'
+export { default as CategoriesSearch } from './CategoriesSearch'
+export { default as StarSearch } from './StarSearch'
+export { default as SubsSearch } from './SubsSearch'
+export { default as BrandsSearch } from './BrandsSearch'
+export { default as ColorSearch } from './ColorSearch'
+export { default as ShippingSearch } from './ShippingSearch'
